requirements.txt
pydantic-ai[openai,anthropic,ag-ui]==0.4.7
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
python-dotenv==1.0.0
httpx>=0.27.2
python-multipart>=0.0.9
pytz==2024.1
mangum==0.17.0